requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.3
awscli==1.18.105
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.1.4
blis==0.4.1
board==1.0
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
combo==0.1.1
configparser==5.0.0
confuse==1.3.0
coverage==5.2
crayons==0.3.1
cryptography==3.1.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
datefinder==0.7.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.1.1
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
funcy==1.14
future==0.18.2
gast==0.3.3
gensim==3.8.3
geographiclib==1.50
geopy==2.0.0
gitdb==4.0.5
GitPython==3.1.8
google==3.0.0
google-api-core==1.22.2
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-automl==2.0.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
googlemaps==4.4.2
grpcio==1.32.0
h5py==2.10.0
helpdev==0.7.1
htmlmin==0.1.12
idna==2.10
image==1.5.32
ImageHash==4.1.0
imageio==2.9.0
imagesize==1.2.0
imutils==0.5.3
inflection==0.5.0
intervaltree==3.0.2
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.4
jupyterlab-server==1.1.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.2.1
kiwisolver==1.2.0
kmodes==0.10.1
lazy-object-proxy==1.4.3
libcst==0.3.10
lightgbm==2.3.1
littleutils==0.2.2
llvmlite==0.33.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mlxtend==0.17.2
modin==0.7.4
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.6
murmurhash==1.0.2
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.6
nested-cv==0.916
networkx==2.4
nltk==3.5
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.0.0
oauthlib==3.1.0
opencv-python==4.4.0.44
openpyxl==3.0.3
opt-einsum==3.3.0
outdated==0.2.0
packaging==20.4
pandas==1.1.2
pandas-flavor==0.2.0
pandas-profiling==2.8.0
pandocfilters==1.4.2
paramiko==2.7.1
parse==1.15.0
parso==0.5.2
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.10.0
pluggy==0.13.1
pprintpp==0.4.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
proto-plus==1.9.1
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
py-spy==0.3.3
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pydub==0.24.1
pyee==7.0.2
pyflakes==2.1.1
Pygments==2.6.1
pyLDAvis==2.1.2
pylint==2.5.3
PyNaCl==1.4.0
pyod==0.8.1
pyparsing==2.4.7
pyppeteer==0.2.2
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyquery==1.4.1
pyrsistent==0.16.0
pytest==5.4.3
pytest-cov==2.10.0
pytest-mock==3.2.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.3
QtPy==1.9.0
Quandl==3.5.0
ray==0.8.6
redis==3.4.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
six==1.14.0
smart-open==2.1.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
suod==0.0.4
tangled-up-in-unicode==0.0.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
visions==0.4.4
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xarray==0.16.0
yarl==1.6.0
zope.interface==5.1.0
